batch: Move file entry storage
Batch storage now owns text-file persistence, while binary and gitlink
persistence live in file-type modules.

Generic entry copy, removal, and readback operations cut across file types.
Keeping them with text storage keeps reset-from-batch tied to the text
persistence module for operations that are not text-specific.

This commit continues narrowing storage by moving generic entry operations
into batch.file_entry_storage and updating reset claim helpers.

Subsequent commits will move tests and boundary checks to the new entry
owner.
